Add engine tests, fix and refactor macroquad code
